+++ src/vite-end.d.ts
@@ -1,3 +1,12 @@
 /// <reference types="vite/client" />
-declare const GITHUB_RUNTIME_PERMANENT_NAME: string
-declare const BASE_KV_SERVICE_URL: string
+
+interface ImportMetaEnv {
+  readonly VITE_COSMOS_ENDPOINT: string
+  readonly VITE_COSMOS_KEY: string
+  readonly VITE_COSMOS_DATABASE: string
+  readonly VITE_COSMOS_CONTAINER: string
+}
+
+interface ImportMeta {
+  readonly env: ImportMetaEnv
+}